+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.1.0] - 07/10/2019
 
 * Improve the snap system for the sheet. Now the sheet only snaps to above snap positions if dragging upwards and snap positions below if dragging downwards
+* Made the background widget (child) optional
   
 ## [0.0.2] - 04/10/2019